検証用プログラムのクレイトを大幅に変更した
