+++ zlearning/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "a-learning"
+name = "zlearning"
 version = "0.1.0"
 authors = ["hidekuno <hidekuno@gmail.com>"]
 edition = "2018"
